Truncate menu help value positions

+++ src/MenuUtil.cpp
@@ -525,9 +525,9 @@ void CMenuPcs::DrawHelpMessageUS(int msgNo, CFont* font, int, int, _GXColor colo
 		strcat(scratch, sMenuUtilSpaceText);
 		font->Draw(scratch);
 
-		float valueX = FLOAT_8033357c + font->GetWidth(scratch) + FLOAT_803335a0;
+		int valueX = static_cast<int>(FLOAT_8033357c + font->GetWidth(scratch) + FLOAT_803335a0);
 		font->SetTlut(1);
-		font->SetPosX(valueX);
+		font->SetPosX(static_cast<float>(valueX));
 		sprintf(scratch, sMenuUtilValueSuffixFormat, *reinterpret_cast<u16*>(itemBase + 6));
 		font->Draw(scratch);
 
@@ -564,8 +564,8 @@ void CMenuPcs::DrawHelpMessageUS(int msgNo, CFont* font, int, int, _GXColor colo
 				}
 
 				int delta = static_cast<int>(*reinterpret_cast<u16*>(itemBase + 6)) - static_cast<int>(currentValue);
-				float deltaX = valueX + font->GetWidth(scratch) + FLOAT_803335a0;
-				font->SetPosX(deltaX);
+				int deltaX = static_cast<int>(static_cast<float>(valueX) + font->GetWidth(scratch) + FLOAT_803335a0);
+				font->SetPosX(static_cast<float>(deltaX));
 				if (delta < 0) {
 					font->SetTlut(3);
 				} else {
